Remove unnecessary return cycle date helpers
TBH, always doubted we needed to keep translating date values to ISO strings before or after interacting with the DB.

My own testing and the unit testing confirms this is the case. So, we can remove the `*AsISO()` helpers entirely. Other than that, we update the helpers to return directly and fix a typo.