Refactor notifications returns period options
https://eaflood.atlassian.net/browse/WATER-4716

As part of the ongoing work to migrate the legacy UI we are replacing the notification journey from the UI and rebuilding in system.

This change refactors the existing logic in an attempt to better align the return options logic with return cycles.

This change also uses the formatLongDate base function to ensure all our dates meet the applications shared expected output