Move return log creation back to its own service
As suspected were moving the creation of the return logs back into its own service. But we do move the generate call into the service as well as the error handling.

Now both the job and ad-hoc processes can taking advantage of this shared service.

Note - we also do a little pre-emptive work ahead of quarterly to reduce the changes needed when we come to amend this engine to cater for it.