Amend `BasePresenter.titleCase()`
We've found that the purpose text we retrieve is all-caps and we want it to be in title case. However `BasePresenter.titleCase()` does not handle this -- it capitalises the first letter of each word while leaving the rest of the word as-is, effectively returning the same all-caps string. Up until now this behaviour was neither desired nor undesired; we simply didn't have a need to convert an all-caps string to title case. Now that we do, we amend `BasePresenter.titleCase()` to first convert the string to lower case before capitalising the first letter of each word, thus forcing all-caps text into title case.