Refactored fetch licence rtn requirements for all
Imagine we have a licence, that has had a changed applied some years back. That means we'll have a number of return cycles to process; multiple summer and all-year ones.

I realised that when we call `FetchLicenceReturnRequirementsService` to args would always be the same. We were not filtering by return cycle ID, only whether the cycle was summer or all-year.

In a change where we are processing more than one summer cycle, for example, we'd be fetching the same requirement data multiple times.

Instead, if we dropped the cycle type filter we could grab all the relevant return cycles for the licence in one go. Then each time we process a cycle we can filter out the matching requirements.

The other benefits are we can move the checks whether to bother doing anything further up in the order, and simplify our tests.