Fix issue with dates in CreateReturnCycleService
It was pointed out that I had messed up the date handling in the CreateReturnCycleService tests. I was dynamically setting the stubbed current date, but was was always asserting against a fixed result.

It would work for the foreseeable future, but next year would most likely break.

Resolved the issue by fixing the 'current date' we use in each test. That way we can better demonstrate what is happening, and assert against fixed/expected values.