Return log for ended licence (#1425)
https://eaflood.atlassian.net/browse/WATER-4547

As part of the work to replace the legacy system NALD we need to generate return logs. 

This PR sets up the logic for voiding and reissuing of return logs when a new return version has been entered. It will use the new dates of the return version to figure out which return logs need to be voided, voids them and then creates new return logs to cover the return cycles from that period forward.