Update model unit tests to use before block
This PR is part of our ongoing work to make unit tests faster and more efficient, focusing on the model unit tests. We are simplifying the database seeding process by moving it to a top-level before block instead of using beforeEach. Since the model unit tests check the relationships between table models, putting all the test data in one place means we only need to seed the database once, instead of before each test. This makes the tests run faster and keeps the code cleaner.