Add /check into bill run setup ready for 2PT supp.
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff bill runs

As part of WATER-4200 we made a series of changes to the bill run setup journey to support Two-part tariff supplementary. What we didn't do is add in the checks to ensure we could create a two-part tariff supplementary.

We've come to do that now and realised two things

- the checks to see if an existing bill run blocks the one being created are confusing
- the end of our bill run setup journey is confusing

We will tackle the first issue in [Refactor bill run setup checks ready for 2PT supp.](https://github.com/DEFRA/water-abstraction-system/pull/1542) This change is resolving the confusion at the end of the journey.

Currently, depending on the bill run type you can be 'ready' to create the bill run at various points in the journey. For example, you could be ready to create a PRESROC two-part tariff after four pages, but an annual in only two!

So, each page had logic that told the controller whether to redirect to the `/create` endpoint or not. When it was called, `/create` would first call `ExistsService`. If it found no blocking bill run, the controller would then call `CreateService`. But if a match was found, `ExistsService` was responsible for generating the page data using `CreatePresenter`, and passing that to the `create.njk` view.

We're breaking our convention used everywhere else that controller, service, presenter and view all line up because they all use the same name.

What we realised when trying to get our heads back into what was happening, was that it's our process that is breaking convention, which is forcing the code to be confusing. In all our other journeys we reach a `/check` page where users can review what they are about to do, or we can highlight issues with the options they have selected. The bill run setup journey is missing a `/check` page!

Adding in a `/check` page not only brings the bill run setup journey inline with others, it means we can better organise our code and also bring it inline with our conventions.