Rename create.njk to check.njk
This page was really the `/exists` page, intended only to display the details of the blocking bill run. As such, we modelled it on pages like `/send` and `/cancel`, using the bill run metadata convention.

Now for most journeys we expect to be replaying the options a user selected. As such, we have followed the convention of the return version's `/check` and used the GOV.UK summary list in its standard form to display them.

Where we have a match, we display its details instead, and expand it to include the date created, and the bill run number.

We've also brought the page inline with `/no-licences`, which was recently added as part of supporting two-part tariff supplementary bill runs. If you select a region that has no licences flagged for two-part tariff supplementary, instead of the `/years` page displaying, you instead see the `/no-licences` page.

It features a 'Return to bill runs` link at the bottom giving users a visible 'get-out'. We've included that in our `/check` page no matter if a match is found or not as a visible means for the user to abort the bill run creation.

Having it and the 'Go to bill run' link clashed. So, instead we've incorporated that into the bill run number, which only gets displayed when their is a blocking match.

Finally, we now have a green confirmation button, so along with replaying their options users get to make an active decision to create a bill run, and no exactly at what point it happens.