Add new submit check service
Now we have a more conventional `/check` endpoint, we can have a 'submit' service that backs a post request to it. It does mean we have to do the exists check again, but that is something we encounter with all check pages; the need to re-check on submission in case of new changes made whilst the user was 'checking'.

But now we have our standard submit pattern; a service that manages the validation, and the results of it, including re-rendering the view in the event of an issue.