Refactor so 'trigger' checks are in ExistsService
This came about when manual testing highlighted the refactored solution was blocking creation of a supplementary bill run when only 1 match is found.

We allow this for supplementary, and it is one of the complexities that makes our checking rules so confusing!

But our refactored logic was just looking at `matches.length`. To make it work, it was going to need to do similar checks to what the `CreateService` was already doing to understand if and what bill run engines could be triggered.

This led us to deciding to move those checks to one place, where the data already sits; the ExistsService!