Fix check page for non-prod edge case
Having moved to using the trigger, when it came to our pesky non-prod edge case we found there was an issue.

It will have a toFinancialYearEnd of 0, so `trigger:` will be 'neither'. But then we had logic assuming 'neither' means we have a match.

It was clearer to move away from the idea of an 'existing bill run', especially as you can still create a bill run with a match in some cases. Instead, we let whether we have populated a value determine if we show it (we'll only populate it _if_ we want to show it!)

So, that just leaves whether we want to show the create bill run button or not. So, we've updated the logic and tests to use to focus on that property instead.