Fix return log helper breaking acceptance tests (#1551)
In [Return log for ended licence](https://github.com/DEFRA/water-abstraction-system/pull/1425), we made some housekeeping changes to the `ReturnLogHelper`. We had gone a bit OTT, passing in dates to `new Date()` to make a new date!

So, we scrubbed most of them, and all unit tests and logic were still okay. However, when we came to run the acceptance tests, we found a bunch were broken.

We tracked it down to the acceptance tests having to pass dates as strings to our `app/services/data/load/load.service.js`. In most cases, that is fine, but `ReturnLogHelper` passes these to `generateReturnLogId()`, which in turn passes them to `formatDateObjectToISO()`.

If they are not a date object, `formatDateObjectToISO()` blows up.

So, this quick fix gets our acceptance tests working again.