Add /check into bill run setup ready for 2PT supp. (#1550)
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff bill runs

As part of [WATER-4200](https://eaflood.atlassian.net/browse/WATER-4200), we made a series of changes to the bill run setup journey to support Two-part tariff supplementary. What we didn't do was add in the checks to ensure we could create a two-part tariff supplementary.

We've come to do that now and realised two things.

- the checks to see if an existing bill run blocks the one being created are confusing
- the end of our bill run setup journey is confusing

We will tackle the first issue in [Refactor bill run setup checks ready for 2PT supp.](https://github.com/DEFRA/water-abstraction-system/pull/1542) This change resolves the confusion at the end of the journey.

Depending on the bill run type, you can be 'ready' to create the bill run at various points in the journey. For example, you could be ready to create a PRESROC two-part tariff after four pages but an annual in only two!

So, each page had logic telling the controller whether to redirect to the `/create` endpoint. When it was called, `/create` would first call `ExistsService`. If it found no blocking bill run, the controller would then call `CreateService`. But if a match was found, `ExistsService` generated the page data using `CreatePresenter` and passed the result to the `create.njk` view.

We're breaking our convention used everywhere else that controller, service, presenter, and view all line up because they all use the same name.

When trying to get our heads back into what was happening, we realised that our process was breaking convention, forcing the code to be confusing. In all our other journeys, we reach a `/check` page where users can review what they are about to do, or we can highlight issues with the options they have selected. The bill run setup journey is missing a `/check` page!

Adding a `/check` page not only aligns the bill run setup journey with others but also allows us to better organise our code and conform to our conventions.