Amend transaction line description for 2PT bills
https://eaflood.atlassian.net/browse/WATER-4725

The business have requested amendments to the transaction line description and associated labels in the transaction files that are generated by the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api).

The labels require an amendment to the **Charging Module API** itself. But the line descriptions are generated by us and sent to the CHA.

This change amends the descriptions we generate for two-part tariff bill runs.

**Annual & Supplementary bill runs**

- `Two-part tariff basic water abstraction charge` to `Two-part tariff first part water abstraction charge`

**Two-part tariff annual and supplementary bill runs**

- `Two-part tariff basic water abstraction charge` to `Two-part tariff second part water abstraction charge`