Housekeeping - stub simple bill run query in tests
DetermineFinancialYearEndService relies on making a simple query to the DB. Currently we are having to create a series of bill run records to support the tests, and then are deleting them afterwards to avoid causing issues for other unit tests.

We can instead refactor the tests to stub the simple query.