Add return period lib functionality (#1546)
* Add return period lib functionality

https://eaflood.atlassian.net/browse/WATER-4716

As part of the ongoing work to migrate the legacy UI we are replacing the notification journey from the UI and rebuilding in system.

This change introduces the logic to handle "quarter" and annual periods.

Quarters are from April to April.

This change also includes functionality to get the current and next return period. This is determined by the due date.