Remove duplicate blocking check
Originally, we were triggered from the legacy code when a bill run was created. Once we took over the create bill run journey this was no longer the case, but the endpoint still remained. So, this meant we needed to still check if our original create bill run endpoint was hit directly.

We've recently removed that, so now the only route to create a bill run is through our setup journey. This means checking again if a bill run is blocked is a duplicate/redundant action.