First refactor - Move annual checks to new service
This is the first of our refactorings. Rather than have services that try to do one thing for all bill run types, we are moving to having a service dedicated to checking all the things for a single bill run type. That way we can see all the conditions for a bill run type in one place.

This is the first service which covers annual, the simplest of the bill run types to checking for any blocks.