Rename Exists to DetermineBlockingBillRunService
Now the checks for each bill run type are being done in separate services, 'Exists' becomes more of an orchestrator. This includes moving out determining the `trigger` functionality to each of those services.

So, rename to matter match its purpose and to link it to the other services we've added.