New FetchLiveBillRun service
This is actually different to the existing (soon to be deleted!) FetchLiveBillRunsService.

By refactoring the bill run checks to their own service it became easier to see that what the supplementary is doing when looking for matches is the same as what FetchLiveBillRunsService was doing.

So, as supplementary no longer needs to call FetchLiveBillRunsService, a lot of its complexity for dealing with supplementary became redundant. Now, it was purely needed for the annual bill run types to double check nothing was in progress.

Plus, as we've dropped the checks from InitiateBillRunService it could be moved into the `/setup folder`.

A rename, a move, and a rewrite all meant we're treating it as a brand new service!