Refactor bill run setup checks ready for 2PT supp. (#1542)
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff bill runs

The first step in supporting two-part tariff bill runs is handling it in the set-up journey; as with annual and supplementary, we have to perform a series of checks regarding the current state of bill runs. For example, you cannot create an annual bill run if one has already been created that year, but you can for supplementary.

We now need to add two-part tariff supplementary into the mix. Our current checks do it as one, the `ExistsService`.

Returning to this area of the code after some time, we've realised it isn't very clear. Yes, it is doing 'one thing'. But that one thing gets very complex very quickly when you start having to cater for the different bill run types and states of the existing bill runs.

So, before we squeeze two-part tariff supplementary in there, we will refactor the process by splitting the work out and putting the checks for each bill run type into their own service.