Add two-part supplementary block check to setup
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff bill runs

In [Add /check into bill run setup ready for 2PT supp.](https://github.com/DEFRA/water-abstraction-system/pull/1550) we gave ourselves a new `/check` page in the journey and the space to better organise all the checks we have to do before deciding whether a bill run can be created.

Then in [Refactor bill run setup checks ready for 2PT supp.](https://github.com/DEFRA/water-abstraction-system/pull/1542) we refactored we switched how we were doing the checks. Rather than services that did one thing for all bill run types, for example, finding matches, we instead added services that did _all_ the checks for a single bill run type. That simplified the logic and allowed us to see in one place the conditions for blocking or allowing a particular bill run to proceed.

We've that tidying up done, we are now ready to support a new bill run type: `two_part_supplementary`.

This change adds the checks needed to confirm if a new `two_part_supplementary` bill run can be created or not.