Add our determine blocking service
The checks for two_part_supplementary are the simplest yet. One charge scheme, and we only have to look at the selected year. And, you are allowed to create multiple two part tariff supplementary bill runs in a given year.

So, all we are really checking is if any bill run is currently in progress (live).