Start moving financial year determination
The final change we need to make is determining what financial year to use.

Like supplementary, you can only create a two-part tariff supplementary for a region and year that has previously had a two-part tariff annual 'sent'.

But unlike supplementary, where the year to run is controlled by us, users pre-select the year to run based on data in the  `licence_supplementary_billing_years` table.

For a given region, if any 2PT licences have been flagged for two-part tariff supplementary, they'll have a record in `licence_supplementary_billing_years`. As part of the setup journey we display the years we find in those records (if no entries we redirect users to the `/no-licences` page).

So, if they, for example, select 2024 and we then find no one has yet run the 2024 two-part tariff annual, we cannot automatically change the year to be 2023 (a feature of normal supplementary bill runs) because there might be no licences flagged for it.

This means we have yet another block of logic we have to squeeze into DetermineFinancialYearEnd, a service that is currently trying to do one thing for all bill run types.

Rather than do that, we are going to move that work into our bill run blocking services, so we can truly say all the conditions for a bill run type are on one place!