Feature returns period lib determine upcoming periods (#1559)
* Add return period lib determine upcoming periods

https://eaflood.atlassian.net/browse/WATER-4716

As part of the ongoing work to migrate the legacy UI we are replacing the notification journey from the UI and rebuilding in system.

This change introduces the logic determine upcoming return periods based on the due date.

Where a due date is the same we want to use the one with the earliest start date by month and date only. 

This is an issue where quarter one and all year have the same due date but the start date for all year is in the previous year. If we used the start date then the all year would always be first. This is a business rule to priorities quarter one over all year.