Refactor returns period presenter to use determine upcoming periods
https://eaflood.atlassian.net/browse/WATER-4716

As part of the ongoing work to migrate the legacy UI we are replacing the notification journey from the UI and rebuilding in system.

This change introduces the logic determine upcoming return periods based on the due date into the presenter.

This logic was previous hard coded in the presenter but was moved out and completed in a previous piece of work - https://github.com/DEFRA/water-abstraction-system/pull/1559