reissue return logs when new return version saved (#1564)
https://eaflood.atlassian.net/browse/WATER-4547

As part of our work to replace NALD in generation of return logs we need to be able to reissue return logs when a new return version is created.

This PR updates the logic for return return requirements to return them from both return versions when there is a changed made. This then enables the voiding and creation of return logs both before and after the new dates that are generated as part of the new return version.