Handle change in end date for a licence (#1538)
https://eaflood.atlassian.net/browse/WATER-4546

As part of the work to generate return logs we need to handle importing a licence from the legacy system NALD. This PR pulls out the logic for identifying which licences have had a end date changed into it's own service. I've then moved the logic for calling the determine billing flags into the import section at the same place as the call to the service to generate any changes to the return logs.