Always pass a year to DetermineBillingPeriods
The comments and code for handling no year being passed to `DetermineBillingPeriods` is redundant. It was there only because when the legacy triggered an Annual bill run, for example, it didn't bother to provide a financial year end.

Now all bill runs are triggered via the setup journey, and it determines the year to use for all bill run types. This means we are always providing a value.

This also meant we could drop a number of the tests.