Update tests for returns period presenter
https://eaflood.atlassian.net/browse/WATER-4716

As part of the ongoing work to migrate the legacy UI we are replacing the notification journey from the UI and rebuilding in system.

This change updates the tests for the returns period presenter to cover al the possible output scenarios.

Previous work was done to use the return periods available in the lib dir. This code calculates all the return periods and provides them sorted by the due date.

The previous work to use this lib is here - https://github.com/DEFRA/water-abstraction-system/pull/1560

This change updated the existing tests, increased them to cover the expected outcome similar to determineUpcomingReturnPeriods tests. But as this is the presenter we want to test the radio button structure for the page render.