Feature returns period submit logic (#1571)
* Feature returns period submit logic
https://eaflood.atlassian.net/browse/WATER-4716

As part of the ongoing work to migrate the legacy UI we are replacing the notification journey from the UI and rebuilding in system.

This change introduces the submit logic for the returns period. It will handle the validation and rendering of an invalid page.

This will not introduce saving the value to session. This will be done in a separate piece of work which will need to initialize the session.