Feature notifications setup returns period session
https://eaflood.atlassian.net/browse/WATER-4716

As part of the ongoing work to migrate the legacy UI we are replacing the notification journey from the UI and rebuilding in system.

This change introduces session save and retrieval for the returns period page. When he user has selected a returns period then this will be saved to the session for the notifications setup journey.