Fix miscalculated preview 2PT charge in review
https://eaflood.atlassian.net/browse/WATER-4849

The business have spotted that when previewing the two-part tariff charge calculation for some licences, that the value returned is more than expected.

They have deduced it is because the preview is incorrectly flagging the licence as having 'public water supply' and therefore is adding the additional £141 charge to the calculation.

This change corrects what data we send to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) to ensure we get present the correct preview charge to users.