Exclude ret reqs from ret vers that have ended (#1580)
https://eaflood.atlassian.net/browse/WATER-4547

As part of the testing of the historic changes, it was noticed that additional return logs were being created that had start dates after their end dates.

After some digging, we figured out that when dealing with a new return version, it was always creating return logs for the entire return cycle. It should only be creating return logs up to the end date of the return version.

This change excludes return requirements from a return version, which have a start date after its end date.