Fix miscalculated preview 2PT charge in review (#1583)
https://eaflood.atlassian.net/browse/WATER-4849

When previewing the two-part tariff charge calculation for some licences, the business spotted that the value returned was more than expected.

They have deduced it is because the preview is incorrectly flagging the licence as having a 'public water supply' and, therefore, adds the additional £141 charge to the calculation.

This change corrects what data we send to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) to ensure we present the correct preview charge to users.