Highlight code with no tests but reported in coverage (#1579)
As part of ongoing improvements we are trying to make sure all code has coverage (where deemed reasonable).

Some of our code may be tested indirectly and this helps cover parts of the codebase.

In this change we are highlighting that the:

- app server does not have/ or need a test
- check endpoints are for testing and is no production code so this is deemed to not need a test
- app plugins are inherently tested when testing with the server running with hapi lab

We can look to test these specifically at a later date.