Remove the legacy import code
https://eaflood.atlassian.net/browse/WATER-4535

This time last year (December 2023) we'd completed a workshop with our ReSP colleagues, and had begun working to a vision that would conclude with NALD being replaced by ReSP. A key part of this would be a complete change in the way we import our licence abstraction (ABS) data.

The legacy process that looks at _all_ licences, processing each one to determine new and changed licences, would be replaced by single licence 'events'. ReSP would provide an event bus from to which our service would be a subscriber. As MVP (because details were scant) we would treat any event as a trigger to process a licence and determine if it is new or changed.

In order to be as 'ready' as possible, we saw an opportunity to replace the existing legacy import and create the basis for that MVP. We would rebuild what the legacy import is doing but refactor the process to be licence specific. For example, the legacy import has a CRM step for all licences, then an abstraction step, then a returns step. We'd refactor that to do the CRM step, then the abstraction step, then the returns step _for a single licence_.

This work was done under [ReSP Import abstraction data prep - water.licence](https://eaflood.atlassian.net/browse/WATER-4535). We got it to a point where the contacts and abstraction steps could be tested, once we'd figured out just how we'd go about verifying it was importing things correctly! We'd then need to tackle returns information.

Piggy-backing on this work were changes need for two-part tariff billing and return version management. These two features needed to know when a licence 'end dates' changed in order to correctly flag it for billing, and in case return logs needed to be reissued.

This was no small job, and got as far as it did thanks to the perseverance of @jonathangoulding . Unfortunately, events overtook this work and a new change in direction has been agreed. Now WRLS itself will replace NALD. This means licences will be created and managed within the service directly. There will be no need for an import process _at all_.

It having changed direction once, there was always the chance it would change back! So, we decided to leave the work done within the app. If nothing else, it's a handy resource to refer to when trying to figure out what the legacy import is actually doing.

But two-part tariff billing and return version management still need to know if a licence's end dates have changed. So, we've added a new job (see [WATER-4728 New 'import' job to trigger return log and supplementary flag checks](https://eaflood.atlassian.net/browse/WATER-4728)) at `/jobs/import-licences` specifically designed to trigger these two processes.

The problem is when looking at the work around, sometimes folks are doing it in this new job. Other times they are updating and amending the legacy licence import replacement we built. There is a concern that something is going to be missed or done incorrectly because the team is getting confused between the two.

So, leaning into the knowledge that source control means deleted code is never truly lost, this change aims to remove the confusion by deleting the replacement legacy import functionality we added.